+++ CHANGELOG.md
@@ -11,6 +11,12 @@
   - All vertices must be coplanar.
   - Zero length segments are not allowed.
   - Self-intersecting segments are not allowed.
+- `Solid`, `Face`, `Vertex`, `Edge`, `HalfEdge`, and `Loop` are now marked `internal`.
+- `Quaternion` is now marked `internal`.
+- `Matrix` is now marked `internal`.
+- `Transform(Matrix matrix)` constructor is now marked `internal`.
+- `SolidConverter` is now marked `internal`.
+- `Elements.Serialization.IFC.ModelExtensions` is now marked `internal`.
 
 ### Added
 - `Identifiable`